Refactor: mws handler 수정
