+++ src/mocks/browser.ts
@@ -5,11 +5,13 @@ import { loginHandlers } from '@/mocks/handler/loginHandlers';
 import { tokenHandlers } from '@/mocks/handler/tokenHandlers';
 import { totalHandlers } from '@/mocks/handler/totalHandlers';
 import { transactionHandlers } from '@/mocks/handler/transactionHandlers';
+import { tokenVerifyHandlers } from '@/mocks/handler/tokenVerifyHandler';
 
 export const worker = setupWorker(
   ...loginHandlers,
   ...tokenHandlers,
   ...authHandlers,
+  ...tokenVerifyHandlers,
   ...emailHandlers,
   ...totalHandlers,
   ...transactionHandlers,
