+++ src/mocks/handler/tokenHandlers.ts
@@ -32,30 +32,4 @@ export const tokenHandlers = [
       { status: 200 },
     );
   }),
-  http.get('*', async ({ request }) => {
-    const authHeader = request.headers.get('authorization');
-
-    if (!authHeader || !authHeader.startsWith('Bearer ')) {
-      return HttpResponse.json({ message: 'Unauthorized' }, { status: 401 });
-    }
-
-    const token = authHeader.replace('Bearer ', '');
-
-    if (!isTokenValid(token)) {
-      return HttpResponse.json({ status: 401, message: '액세스 토큰이 만료되었습니다.' }, { status: 401 });
-    }
-  }),
-  http.post('*', async ({ request }) => {
-    const authHeader = request.headers.get('authorization');
-
-    if (!authHeader || !authHeader.startsWith('Bearer ')) {
-      return HttpResponse.json({ message: 'Unauthorized' }, { status: 401 });
-    }
-
-    const token = authHeader.replace('Bearer ', '');
-
-    if (!isTokenValid(token)) {
-      return HttpResponse.json({ status: 401, message: '액세스 토큰이 만료되었습니다.' }, { status: 401 });
-    }
-  }),
 ];
